ecommerce-pipeline/requirements.txt
# Add Python packages here, one per line. Pin versions where possible.


dbt-snowflake
apache-airflow-providers-amazon